requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
requests==2.31.0
psycopg2-binary==2.9.9
geopy==2.4.1
boto3==1.34.0
botocore
s3transfer>=0.9.0
celery
redis
python-dotenv
requests-aws4auth
beautifulsoup4
supabase>=2.0.0
pgvector>=0.2.0
reductoai>=0.1.0
Pillow>=10.0.0
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.2.11
langchain-openai>=0.1.17
langchain-anthropic>=0.1.0
langgraph>=0.2.5
langgraph-checkpoint>=2.1.0,<3.0.0
langgraph-checkpoint-postgres>=2.0.0,<3.0.0
psycopg>=3.1.0
psycopg-pool>=3.1.0
pydantic-settings>=2.3.4
sentence-transformers>=2.2.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
cohere>=4.0.0
voyageai>=0.3.5
pymupdf>=1.24.0
python-docx>=1.1.0
